+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1
+
+* Customizable radius
+
 ## 0.1.0
 
-* Implement `RoundedBackgroundText` and `RoundedBackgroundTextField`
+* Implement `RoundedBackgroundText`
+* Implemnet `RoundedBackgroundTextField`
+* Implemnet `RoundedBackgroundTextSpan`